Algoritmos de consenso ya funcionan

+++ agregador/src/main/java/aplicacion/domain/algoritmos/AlgoritmoConsenso.java
@@ -31,5 +31,5 @@ public abstract class AlgoritmoConsenso {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    public abstract List<Hecho> curarHechos(Map<Hecho, Integer> cantidadPorHecho, Integer totalFuentes);
+    public abstract List<Hecho> curarHechos(Map<Hecho, Long> cantidadPorHecho, Long totalFuentes);
 }
